pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imgtxt"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0",
    "typer>=0.9",
    "wcwidth>=0.2",
]

[project.scripts]
imgtxt = "imgtxt.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["imgtxt"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["imgtxt"]

[tool.coverage.report]
fail_under = 80
show_missing = true
